Rename filter_class to filterset_class
In Django REST Framework and django-filters, filter_class was the original attribute name for specifying a FilterSet on a ViewSet. However, this was deprecated and
replaced with filterset_class in newer Django versions to be more explicit and consistent with DRF naming conventions.

This fixes the flaky test_list_panels_filter_types unit test.