Update the evaluation's last_updated timestamp when adding a review
Weirdly, evaluations have both a `last_updated` *and* a `modified`
timestamp. `modified` gets automatically updated by Django, but
`last_updated` needs to be updated manually.

This fixes the odd behavior that `last_updated` (which is also used
in the UI to render review dates) wasn't updated when a new review
was added.